Time: 11 ms (62.14%) | Memory: 4 MB (61.43%) - LeetSync

+++ 883-car-fleet/car-fleet.rs
@@ -5,9 +5,9 @@ impl Solution {
         let mut cars = Vec::from_iter(position.into_iter().zip(speed));
         cars.sort();
         let mut count = 0;
-        let mut time = 0f64;
+        let mut time = 0f32;
         for car in cars.iter().rev() {
-            let mut curr_time: f64 = (target - car.0) as f64 / car.1 as f64;
+            let mut curr_time: f32 = (target - car.0) as f32 / car.1 as f32;
             if time < curr_time{
                 count = count + 1;            
                 time = curr_time;
